% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesMclust.R
\name{bayesMclust}
\alias{bayesMclust}
\title{Bayesian Mclust}
\usage{
bayesMclust(
  X,
  R,
  thin,
  K,
  alpha = 1,
  dataType = "G",
  method = "complete",
  metric = "euclidean"
)
}
\arguments{
\item{X}{Data to cluster as a matrix (items to cluster in rows).}

\item{R}{The number of iterations in the sampler.}

\item{thin}{The factor by which the samples generated are thinned, e.g. ig ``thin=50`` only every 50th sample is kept.}

\item{K}{A vector of the number of components to include (the upper bound on the number of clusters found) in each model.}

\item{alpha}{The concentration parameter for the stick-breaking prior and the weights in the model.}

\item{method}{The linkage used in ``hclust``.}

\item{metric}{The measure of distance used in creating a distance matrix for ``hclust``.}
}
\value{
Matrix of MCMC samples generated (each row corresponds to a different sample).
}
\description{
A range of Bayesian mixture model initialised using ``hclust`` with a cut for each value of the vector ``K`` given.
Named after the ``mclust::Mclust`` function as it follows a similar logic.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# Number of components
K_max <- 15

# MCMC samples
m <- bayesMclust(X, R, thin, K_max = K_max)

library(ggplot2)
library(magrittr)
m$BIC \%>\%
  ggplot(aes(x = K, y = BIC, colour = K)) +
  geom_boxplot()

psmKbest <- m$sample_list[[m$bestModel]]$samples \%>\%
  createSimilarityMat()
}
