\name{ccbm-package}
\alias{ccbm-package}
\alias{ccbm}
\docType{package}
\title{
\packageTitle{ccbm}
}
\description{
\packageDescription{ccbm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ccbm}
\packageIndices{ccbm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ccbm}

Maintainer: \packageMaintainer{ccbm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
