% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{samplerFactory}
\alias{samplerFactory}
\title{Factory for different sampler subtypes.}
\description{
The factory allows the type of mixture implemented to change 
based upon the user input.
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor \itemize{
\item Parameter: samplerType - the density type to be modelled
\item Parameter: K - the number of components to model
\item Parameter: labels - the initial clustering of the data
\item Parameter: concentration - the vector for the prior concentration of 
the Dirichlet distribution of the component weights
\item Parameter: X - the data to model
}}
}}

