% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleSemisupervisedMixtureModel}
\alias{sampleSemisupervisedMixtureModel}
\title{Mixture model}
\usage{
sampleSemisupervisedMixtureModel(
  X,
  K,
  labels,
  fixed,
  dataType,
  R,
  thin,
  concentration,
  seed
)
}
\arguments{
\item{X}{The data matrix to perform clustering upon (items to cluster in rows).}

\item{K}{The number of components to model (upper limit on the number of clusters found).}

\item{labels}{Vector item labels to initialise from.}

\item{fixed}{Binary vector of the items that are fixed in their initial label.}

\item{dataType}{Int, 0: independent Gaussians, 1: Multivariate normal, or 2: Categorical distributions.}

\item{R}{The number of iterations to run for.}

\item{thin}{thinning factor for samples recorded.}

\item{concentration}{Vector of concentrations for mixture weights (recommended to be symmetric).}
}
\value{
Named list of the matrix of MCMC samples generated (each row 
corresponds to a different sample) and BIC for each saved iteration.
}
\description{
Performs MCMC sampling for a mixture model.
}
