% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semisupervisedMixtureModel.R
\name{semisupervisedMixtureModel}
\alias{semisupervisedMixtureModel}
\title{Gaussian mixture model}
\usage{
semisupervisedMixtureModel(
  X,
  R,
  thin,
  initial_labels,
  fixed,
  K_max = 50,
  alpha = 1,
  dataType = 0,
  seed = sample.int(.Machine$integer.max, 1)
)
}
\arguments{
\item{X}{Data to cluster as a matrix (items to cluster in rows).}

\item{R}{The number of iterations in the sampler.}

\item{thin}{The factor by which the samples generated are thinned, e.g. ig ``thin=50`` only every 50th sample is kept.}

\item{initial_labels}{Labels to begin from.}

\item{fixed}{Binary vector indicating which items are fixed in their initial label.}

\item{K_max}{The number of components to include (the upper bound on the number of clusters found).}

\item{alpha}{The concentration parameter for the stick-breaking prior and the weights in the model.}
}
\value{
Named list of the matrix of MCMC samples generated (each row
corresponds to a different sample) and BIC for each saved iteration.
}
\description{
A Bayesian mixture model using independent Gaussians. The priors
are empirical and follow the suggestions of Richardson and Green <https://doi.org/10.1111/1467-9868.00095>.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples and BIC vector
samples <- gaussianMixtureModel(X, R, thin)

# Predicted clustering and PSM
pred_cl <- mcclust::maxpear(samples$samples)
psm <- createSimilarityMatrix(pred_cl)
}
