% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tAdjustedSampler}
\alias{tAdjustedSampler}
\title{Base class for adding a t-distribution to sweep up outliers in the 
model.}
\description{
The class that the specific TAGM types inherit from.
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor \itemize{
\item Parameter: K - the number of components to model
\item Parameter: labels - the initial clustering of the data
\item Parameter: concentration - the vector for the prior concentration of 
the Dirichlet distribution of the component weights
\item Parameter: X - the data to model
}}

\item{\code{updateOutlierWeights}}{Updates the weight of the outlier component.}

\item{\code{updateWeights}}{Update the weights of each component based on current 
clustering, excluding the outliers.}

\item{\code{sampleOutlier}}{Sample is the nth item is an outlier. \itemize{
\item Parameter n: the index of the individual in the data matrix and 
allocation vector.
}}

\item{\code{updateAllocation}}{Sample a new clustering and simultaneously allocate 
items to the outlier distribution.}

\item{\code{calcTdistnLikelihood}}{Virtual placeholder for the function that calculates 
the likelihood of a given point in a t-distribution. \itemize{
\item Parameter: point - a data point.
}}
}}

