% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tagmGaussian}
\alias{tagmGaussian}
\title{T-ADjusted Gaussian Mixture (TAGM) type}
\description{
The sampler for the TAGM mixture model with assumption of 
independent features.
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor \itemize{
\item Parameter: K - the number of components to model
\item Parameter: labels - the initial clustering of the data
\item Parameter: concentration - the vector for the prior concentration of 
the Dirichlet distribution of the component weights
\item Parameter: X - the data to model
}}

\item{\code{printType}}{Print the sampler type called.}

\item{\code{calcBIC}}{Calculate the BIC of the model.}

\item{\code{calcTdistnLikelihood}}{Calculate the likelihood of a given data point 
the gloabl t-distribution. \itemize{
\item Parameter: point - a data point.
}}
}}

