% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusClustering.R
\name{consensusClustering}
\alias{consensusClustering}
\title{Consensus clustering}
\usage{
consensusClustering(
  X,
  S,
  R,
  initial_labels = NULL,
  K_max = 50,
  alpha = 1,
  dataType = "G"
)
}
\arguments{
\item{X}{Data to cluster as a matrix (items to cluster in rows).}

\item{S}{The size of the ensemble.}

\item{R}{The chain depth in each learner.}

\item{initial_labels}{Labels to begin from (if ``NULL`` defaults to a stick-breaking prior).}

\item{K_max}{The number of components to include (the upper bound on the number of clusters found).}

\item{alpha}{The concentration parameter for the stick-breaking prior and the weights in the model.}
}
\value{
matrix of samples generated from each learner (each row corresponds to a different sample).
}
\description{
Consensus clustering of Gaussian mixture models.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Ensemble parameters
S <- 100
R <- 100

# Samples from each model
samples <- consensusClustering(X, S, R)

# Predicted clustering and consensus matrix
pred_cl <- mcclust::maxpear(samples)
cm <- createSimilarityMatrix(pred_cl)
}
