% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mixtureModel}
\alias{mixtureModel}
\title{Mixture model}
\usage{
mixtureModel(X, K, labels, dataType, R, thin, concentration)
}
\arguments{
\item{X}{The data matrix to perform clustering upon (items to cluster in rows).}

\item{K}{The number of components to model (upper limit on the number of clusters found).}

\item{labels}{Vector item labels to initialise from.}

\item{dataType}{String, "G", "MVN" or "C" for independent Gaussians, Multivariate Normal or Categorical distributions.}

\item{R}{The number of iterations to run for.}

\item{thin}{thinning factor for samples recorded.}

\item{concentration}{Vector of concentrations for mixture weights (recommended to be symmetric).}
}
\value{
Matrix of MCMC samples.
}
\description{
Performs MCMC sampling for a mixture model.
}
